allow net-ldap 0.4.x
the current release of net-ldap is broken in jruby. specifically, it has some character weirdness in its gemspec which causes it to not install. this issue is fixed in its master branch, but the version has been bumped to 0.4.0.

This commit allows us to use the current net-ldap from github.