Add additional_ldap_filter suport.
I found myself having to filter out certain users during the Initial
user lookup in LDAP. This is because for some attributes like "mail" we
had duplicate user entries in our directory of over 30k Users.
This commit adds the capability to add a RFC2254 formatted String as an
optional additional filter.
This is joined using (&) to the default filter constructed.