Check if user is present in any listed group
Target #149.

Should not be a breaking change since both the code and the documentation
clearly state that `required_groups` members either are strings,
or must have only two elements when any member is an array.

The feature here is that we can list a number of groups after the attribute
and at least any of them has to be matched to be able to pass the verification.